requirements.txt
Flask==3.0.3
requests==2.32.3
apscheduler==3.10.4
pymysql==1.1.1
python-dotenv==1.0.1